refactor(git): extract shared git-token-refresh helper for ECS runtimes
